Permission checking integration

+++ src/php/api.follow.php
@@ -24,6 +24,11 @@ class Follow extends Core {
     // search users
     
     function get_search($keyword = '') {
+		// Check permissions
+		if(!$this->permission->check()) {
+			return $this->permission->errorMessage();
+		};
+		
 		$return = array();
 		
 		$limit = 10;
@@ -64,6 +69,11 @@ class Follow extends Core {
 	
     // self only
     function get_suggestions($ref_bool = NULL, $keyword = '') {
+		// Check permissions
+		if(!$this->permission->check()) {
+			return $this->permission->errorMessage();
+		};
+		
 		$return = array();
 		
 		$limit = 10;
@@ -111,6 +121,11 @@ class Follow extends Core {
 	}
 	
 	function get_referral() {
+		// Check permissions
+		if(!$this->permission->check()) {
+			return $this->permission->errorMessage();
+		};
+		
 		$return = array();
 		
 		$query = "SELECT RU.company_ID, RU.user_ID, RU.user_name, CONCAT(RU.user_name_first, ' ', RU.user_name_last) AS name" // , GROUP_CONCAT(UF.group_ID) AS groups
@@ -132,6 +147,11 @@ class Follow extends Core {
 	}
 	
 	function get_referrals() {
+		// Check permissions
+		if(!$this->permission->check()) {
+			return $this->permission->errorMessage();
+		};
+
 		$return = array();
 		
 		$query = "SELECT U.company_ID, U.user_ID, U.user_name, CONCAT(U.user_name_first, ' ', U.user_name_last) AS name" // , GROUP_CONCAT(UF.group_ID) AS groups
@@ -157,6 +177,12 @@ class Follow extends Core {
 		$user_ID = preg_replace( '/[^0-9]+/', '', $user_ID);
 		if (!$company_ID) $company_ID = COMPANY_ID;
 		if (!$user_ID) $user_ID = USER_ID;
+		
+		// Check permissions
+		if(!$this->permission->check(array("user_ID" => $user_ID, "company_ID" => $company_ID))) {
+			return $this->permission->errorMessage();
+		};
+		
 		$return = array();
 
 		$query = "SELECT U.user_ID, U.company_ID, user_name, CONCAT(user_name_first, ' ', user_name_last) AS name, UF.timestamp as following, GROUP_CONCAT(UF.group_ID) AS groups" // , UF.group_ID
@@ -185,6 +211,12 @@ class Follow extends Core {
 		$user_ID = preg_replace( '/[^0-9]+/', '', $user_ID);
 		if (!$company_ID) $company_ID = COMPANY_ID;
 		if (!$user_ID) $user_ID = USER_ID;
+		
+		// Check permissions
+		if(!$this->permission->check(array("user_ID" => $user_ID, "company_ID" => $company_ID))) {
+			return $this->permission->errorMessage();
+		};
+		
 		$return = array();
 		
 		$query = "SELECT U.user_ID, U.company_ID, user_name, CONCAT(user_name_first, ' ', user_name_last) AS name, GROUP_CONCAT(group_ID) AS groups" //
@@ -212,6 +244,12 @@ class Follow extends Core {
 		$user_ID = preg_replace( '/[^0-9]+/', '', $user_ID);
 		if (!$company_ID) $company_ID = COMPANY_ID;
 		if (!$user_ID) $user_ID = USER_ID;
+		
+		// Check permissions
+		if(!$this->permission->check(array("user_ID" => $user_ID, "company_ID" => $company_ID))) {
+			return $this->permission->errorMessage();
+		};
+		
 		$return = array();
 		
 		$query = "SELECT U.user_ID, U.company_ID, user_name, CONCAT(user_name_first, ' ', user_name_last) AS name, UF.timestamp as following, GROUP_CONCAT(UF.group_ID) AS groups" // , UF.group_ID
@@ -240,7 +278,12 @@ class Follow extends Core {
 	function get_group($group_ID=NULL) { // $type='user'
 		$group_ID = preg_replace( '/[^0-9]+/', '', $group_ID);
 		$return = array();
-
+		
+		// Check permissions
+		if(!$this->permission->check(array("group_ID" => $group_ID))) {
+			return $this->permission->errorMessage();
+		};
+		
 		$db_where = array('user_ID' => USER_ID);
 		if ($group_ID != '') {
 			$db_where['group_ID'] = $group_ID;
@@ -261,6 +304,11 @@ class Follow extends Core {
 
 	// make group
 	function post_group($request_data=NULL) { // $type='user'
+		// Check permissions
+		if(!$this->permission->check($request_data)) {
+			return $this->permission->errorMessage();
+		};
+		
 		$insert = array(
 			'user_ID' => USER_ID,
 			'group_name' => $request_data['group_name'],
@@ -275,7 +323,12 @@ class Follow extends Core {
 	function delete_group($group_ID=NULL) {
 		$group_ID = preg_replace( '/[^0-9]+/', '', $group_ID);
 		if (!$group_ID || $group_ID < 0) return;
-
+		
+		// Check permissions
+		if(!$this->permission->check(array("group_ID" => $group_ID))) {
+			return $this->permission->errorMessage();
+		};
+		
 		$this->db->delete('follow_groups', array('group_ID' => $group_ID));
 		$this->db->delete($this->table, array('group_ID' => $group_ID));
 
@@ -288,7 +341,11 @@ class Follow extends Core {
 		$user_ID = preg_replace( '/[^0-9]+/', '', $user_ID);
 		if (!$company_ID && !$user_ID) return;
 		
-
+		// Check permissions
+		if(!$this->permission->check(array("user_ID" => $user_ID, "company_ID" => $company_ID))) {
+			return $this->permission->errorMessage();
+		};
+		
 		$return = array(
 			'user_ID' => (int) $user_ID,
 			'company_ID' => (int) $company_ID,
@@ -337,7 +394,12 @@ class Follow extends Core {
 		$user_ID = preg_replace( '/[^0-9]+/', '', $user_ID);
 		$group_ID = preg_replace( '/[^0-9]+/', '', $group_ID);
 		if (!$company_ID && !$user_ID) return;
-
+		
+		// Check permissions
+		if(!$this->permission->check(array("user_ID" => $user_ID, "company_ID" => $company_ID))) {
+			return $this->permission->errorMessage();
+		};
+		
 		$insert = array(
 			'user_ID' => USER_ID,
 			'company_ID' => COMPANY_ID,
@@ -360,7 +422,12 @@ class Follow extends Core {
 		$user_ID = preg_replace( '/[^0-9]+/', '', $user_ID);
 		$group_ID = preg_replace( '/[^0-9]+/', '', $group_ID);
 		if (!$company_ID && !$user_ID) return;
-
+		
+		// Check permissions
+		if(!$this->permission->check(array("user_ID" => $user_ID, "company_ID" => $company_ID))) {
+			return $this->permission->errorMessage();
+		};
+		
 		if ($group_ID) {
 			$this->db->delete($this->table, array(
 				'user_ID' => USER_ID,
